Introduce window-producing graphics library
Starting with just the glutin-based window production, and the gl
components will happen in coming commits. The window is hard coded to be
600x600, and the only input it operates on is a window close, in which
case the main loop in the main breaks.

Adding the App and Window structs (instead of just one) as I know that I
will need vertex and shader related structs in the App. Creating the App
now (with one member currently), to prepare for this incoming use case.

This uses glutin: https://github.com/tomaka/glutin to create the window,
currently only using it for the "physical" window and input.

This may be parameterized more in the future, making the glutin::Window
public, but constraining it now to keep the interface as small and
easy-to-understand as possible.