Introduce Rect struct, substruct of VertexData
This introduces a Rect struct, which holds a subset of what the
VertexData help, namely the dimensions and location of the rectangle
that was encoded directly into the VertexData. Although this does not
add new functionality, it will make it easier to see the upcoming change
to allow for a list of Rects in the VertexData.

Note that this fixes a naming bug with the width/height x/y, where they
were being improperly named in the app's constructor, and then swapped
in the VertexData's constructor.