Accept final rustfmt and include diagrams
These were the final pieces of code that I felt the comment and line
breaks helped with conveying what purpose the data served, so including
comments to clarify. This has actually lead me to notice that the
elements should come from the Rect as well, it is only with the Rect's
data that the 6 elements work as we want.