Rename VertexData to VertexBuffers
This struct does involve itself almost entirely with the buffer data,
the only non-buffer related pieces being the Rect and the vertex_width.
The Rect should really hold the vertex_width, as that struct evolves to
be more rich, it will contain it. The (now) VertexBuffers struct holds
the buffers for OpenGL, and the struct that fills it, so naming it as
such to help convey what it does better.