Specify rect function and cycle-time
This allows us to move rectangles around the screen on independent
reset-to-the-beginning cycles. Things are only allows to move linearly.

We expose a rectangle-updating function on the App, and an update on the
rectangle itself. We use a new original x and y for this use case, to
know where we should offset from.

We move one cross faster than the other to show their independent cycle
times.