Hold vec of VertexSpecable, over only Rect structs
This will allow for other shapes to be drawn, as long as they can
provide the vertex information, and be told to update their position.

We need to place these in a Box, as traits are not size-able at compile
time, so we are introducing heap allocations, will look into if that
will be a problem, but seems to run as smoothly now with the larger of
these, admittedly small, scenes.