Reorder vertex structs
This presents them in a way that will hopefully be more readable.

VertexData is followed by VertexSpecable, of which it holds a Vec of.
VertexSpecification follows, which is part of one of the function
signatures of VertexSpecable. Then color and Location are introduced
right before the Rect that uses them, each of them followed by their
impls. The only exception being VertexData, whose impls come last, as it
pulls everything together.