Introduce Triangles as VertexSpecable
These will be used in the coming commit as a valid shape source. We can
see that they are always the width and height specified, with the point
end at the top. They are a single triangle as far as the drawing is
concerned, composed of exactly 3 points (so surprise as a triangle is
defined as 3 points, but we do have to use the element types to keep
consistent with VertexSpecable which may be multiple triangles, but want
to reuse points).