Introduce Vertex and ElementTriangle structs
These are useful in letting the logic that defines a shape and how it is
represented think about each Vertex, and then the elements that define
each one. It changes no representation as far as what is passed to the
GL calls, but does involves some changes in how we represent it
before it gets to the GL calls.

Instead of a vector of vertex data, that we must know is width 5,
instead we now have a struct that represents one vertex, with the 5
fields in it.  We also now have a more sane way to track the elements,
which must always be sets of 3, a triangle represented per set of
elements.

This does involve transforming these easier-to-reason about structs into
the vectors that they must eventually transform to when we get to the GL
calls. The two things that make this not just a mechanical
transformation are:
 * we need to be able to add a single number to all members of the
   ElementTriangle, added with `add_vertex_offset`
 * we don't need to divide by the vertex_width anymore when calculating
   the vertex offset to offset the elements by taking into consideration
   all previous vertices that have been added