Accept VertexSpecables for drawing
Instead of the VertexData holding the vertices to draw, instead accept
them on drawing. This will let us define a cleaner interface, removing
the need to update the shapes that the VertexData holds. This will also
let us draw different numbers of shapes, if certain things appear and
disappear on screen.