Decouple VertexSpecable and Updateable
This means that the vertex logic only has to worry about accepting
things that know how to turn into vertex information, and the logic of
updating is something specific to the shapes that want to deal with
updating their position.

This does involved some different use of these traits, where the
functions that wanted a VertexSpecable, really only wanted to bound
whatever was coming in as VertexSpecable. We have a new type
UpdateSpecable which is just a combination of VertexSpecable and
Updateable. We also have to specify that we don't need the V to be sized
in our bounding of V (the VertexSpecable).