Initial mandelbrot divergence function
Naming the file eq as it is the equation for mandelbrot viewing.

This uses num to get the complex numbers, which are crucial to
calculating if a complex number is in the mandelbrot set.

I found

https://en.wikipedia.org/wiki/Mandelbrot_set#Formal_definition

very useful in getting to realize how the mandelbrot set works.

This implementation is binary in what it shows, a number is either in
the mandelbrot set or it is not, once it falls greater than 2, it
immediately pops out. The float returned should be more useful in
determining how divergent a given value is in the future. This has a
hard coded 25 iteration max, which I am also likely to play with in the
future.