Generate png for mandelbrot set
The concept the Frame introduces is that given a screen definition (it's
width and height) and x and y ranges, we can map a given pixel to where
it represents in the x and y coords. Using this and the image library we
are able to determine whether a pixel should be filled in or not based
on its divergence from the function of the previous commit. The pieces
likely to change here in the future are determining how colorful to make
something based on its divergence, with the pixel construction in
`gen_pixel` being what will change the most. We offer the ability to get
an ImageRgba8 and do with it what you will, or write the image to a hard
coded file, this will be parameterized soon.