Accept filename prefix in write_png
This allows us to have different output png files for different input
csv files.

Going through the exercise of not risking panicking, by returning an
error if this errors and letting it fall through to a higher call. This
will contain little information of what went wrong, but if the frame
construction is able to operate on the path, the happy path at least
will not fall through to this error logic.