Introduce Renderer struct
This holds a subset of what the App was providing, the shaders and the
vertex data. I am intending to allow a Vector of these to exist, some
shaders and vertices will be with textures, some can be with colors,
and potentially more in the future. This also keeps the caller of the
App::new from needing to know which shources go with what vertex
layouts. Want to continue looking into if I can keep some shaders from
only compiling with certain vertices.