Refactor vertex attribute layout
The color and position were never tied to the individual shader sources,
so making them independent of those entities. Instead we now have a
RenderingPipelineSource which knows of the
 * shader glsl sources
 * all vertex attributes that go into the glsl files
 * the width of a single vertex

Would still like to somehow at compile time know to only allow vertex
specifications of the same width that the shaders want, but leaving that
untouched for now.