Isolate knowledge of mapping glsl to shaders
Previously the caller of the shader building would need to know whether
it was a vertex shader or fragment shader, now that mapping does not
need to escape the shader/shader_source modules.

This does involve changing the enum that was the shader types, instead
to a trait; and I feel this does flow more smoothly now.