Introduce Texture vertex setup and layout
This will allow us to set up textures of u8 RBGA colors to map to shapes
we render. There will be work to allow multiple textures later, and I
still really would like to restrict certain vertices to certain vertex
layouts, as it is possible to pass color vertices to a texture shader
and the other way around. May parameterize the RBGA color definition and
u8 pixel color representation, but staying with simple behavior now.

This does involve the loading of the texture related pieces, which is
disjoint from the shader source logic. Accepting the 3 values of width
and height and data from the caller of the texture construction, which
allows for any texture of u8s to be mapped.