Introduce app mode of mandelbrotviewer
This allows you to view the mandelbrot set in an opengl app using the
texture functionality introduced before.

Updating the help text to reflect the new way to call this binary.

We have a single Rectangle that we draw from 2 triangles, with a texture
loaded that displays the mandelbrot set.