Introduce Actions in window module
The existing boolean to represent whether to close the app is introduced
here as Closed, and we also have the Resized Action. I think it will be
more beneficial to only re-export what I want from the glutin Event.
There may be a problem with handling multiple events in one poll_events
call, will address that in the future.