Introduce binary that can view images
This uses the texture loading introduced previously, and some new logic
to handle not stretching the image if the dimensions of the screen do
not match the dimensions of the image.

The png image must be u8 as opposed to u16, and you must supply the
filename to read when calling the program.