Introduce SI-Unit capabilities
The basic idea with these units is that if you have a measurement, you
can add and subtract measurements of the same unit, and when you
multiply/divide measurements by each other, you get a measurement of a
new unit, derived from the multiplication/division.

e.g.

    meter(5) + meter(6) = 11 m
    meter(3) + gram(3) does not compile
    meter(6) / second(3) = 2 m/s
    meter(6) * second(3) = 18 m^2

This is accomplished with type-level integers, keeping track of what
power each unit is to. This only works for the base 7 si units that we
have implemented here. This is heavily derived from
https://github.com/Boddlnagg/units.